fix: 데이터 파일 통합 (#19)
